Move CallChain and PrimitiveClient into their own modules
